Refactor: move `githubGet` helper for re-use
